Share ResourceReaderInstance
